[Simulator-Modification] findout that there is a special class (base class) that enable simulation without GUI 'tgSimView' and for the GUI 'tgSimViewGraphics'; integrate the rendering flag with the command line parser from boost for better using instead of building the project to change the rendering flag

+++ src/core/tgSimulation.cpp
@@ -269,7 +269,7 @@ void tgSimulation::run() const
     m_view.run();
 }
 
-void tgSimulation::run(int steps) const
+void tgSimulation::run(long long steps) const
 {    
     m_view.run(steps);
 }
